@@ -1,27 +1,19 @@
-import logo from './logo.svg';
-import './App.css';
-// import { Outlet } from 'react-router-dom'
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import Home from './pages/home/Home'
+import Login from './pages/login/Login'
+
 
 function App() {
-    return (
-        <div className="App">
-            <header className="App-header">
-                <img src={logo} className="App-logo" alt="logo" />
-                <p>
-                    Edit <code>src/App.js</code> and save to reload.
-                </p>
-                <a
-                    className="App-link"
-                    href="https://reactjs.org"
-                    target="_blank"
-                    rel="noopener noreferrer"
-                >
-                    Learn React
-                </a>
-            </header>
-            {/* <Outlet /> */}
-        </div>
-    );
+	return (
+		<div className="App">
+			<BrowserRouter>
+				<Routes>
+					<Route path="/" index element={ <Home /> } />
+					<Route path="login" element={ <Login /> } />
+				</Routes>
+			</BrowserRouter>
+		</div>
+	);
 }
 
 export default App;
